GenOL/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
absl-py==2.1.0
accelerate==0.33.0
bitsandbytes==0.43.3
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
easydict==1.13
faiss-cpu==1.8.0.post1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.3
grpcio==1.65.5
huggingface-hub==0.24.6
idna==3.7
jinja2==3.1.4
kiwisolver==1.4.5
kornia==0.7.3
markdown==3.7
markupsafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
onedrivedownloader==1.1.3
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
pip==24.2
protobuf==5.27.3
psutil==6.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pyparsing==3.1.2
pytz==2024.1
pyyaml==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
setuptools==72.1.0
sympy
tensorboard
timm
tokenizer
torch==2.4.0
torch-optimizer
torchvision
tqdm==4.66.5
transformers==4.40.0
diffusers